rename config.js to config.ts
