+++ src/services/Youtube.js
@@ -1,4 +1,4 @@
-import config from '../config.js'
+import config from '../config'
 import { getJson, asQueryArgs } from '../net/xhr.js'
 
 const get = async (url, args) => {
